Move utils to ddyaml
