+++ datadog_checks_base/datadog_checks/base/__init__.py
@@ -5,6 +5,7 @@ from .__about__ import __version__
 from .checks import AgentCheck
 from .checks.openmetrics import OpenMetricsBaseCheck
 from .config import is_affirmative
+from .ddyaml import yaml_load_all_force_loader, yaml_load_force_loader
 from .errors import ConfigurationError
 from .utils.common import ensure_bytes, ensure_unicode, to_native_string, to_string
 
@@ -39,4 +40,6 @@ __all__ = [
     'is_affirmative',
     'to_native_string',
     'to_string',  # For backwards compat (was renamed to `to_native_string`).
+    'yaml_load_force_loader',
+    'yaml_load_all_force_loader',
 ]
