+++ datadog_checks_base/datadog_checks/base/ddyaml.py
@@ -114,6 +114,26 @@ def get_stream_name(stream):
         return "<file>"
 
 
+def yaml_load_force_loader(stream, Loader):
+    """ Override the default monkey patch for this call """
+    log.debug(
+        "`%s` YAML loader is used instead of the default one, please make sure it is safe to do so", Loader.__name__
+    )
+    if pyyaml_load is None:
+        return yaml.load(stream, Loader)
+    return pyyaml_load(stream, Loader)
+
+
+def yaml_load_all_force_loader(stream, Loader):
+    """ Override the default monkey patch for this call """
+    log.debug(
+        "`%s` YAML loader is used instead of the default one, please make sure it is safe to do so", Loader.__name__
+    )
+    if pyyaml_load_all is None:
+        return yaml.load_all(stream, Loader)
+    return pyyaml_load_all(stream, Loader)
+
+
 def monkey_patch_pyyaml():
     global pyyaml_load
     global pyyaml_load_all
